fix(rust): List<null> chunked builder should take care of series name
