+++ crates/polars-core/src/chunked_array/builder/list/boolean.rs
@@ -69,26 +69,3 @@ impl ListBuilderTrait for ListBooleanChunkedBuilder {
         self.fast_explode
     }
 }
-
-impl ListBuilderTrait for LargeListNullBuilder {
-    #[inline]
-    fn append_series(&mut self, _s: &Series) -> PolarsResult<()> {
-        self.push_null();
-        Ok(())
-    }
-
-    #[inline]
-    fn append_null(&mut self) {
-        self.push_null()
-    }
-
-    fn finish(&mut self) -> ListChunked {
-        unsafe {
-            ListChunked::from_chunks_and_dtype_unchecked(
-                "",
-                vec![self.as_box()],
-                DataType::List(Box::new(DataType::Null)),
-            )
-        }
-    }
-}
